+++ src/main/java/rvn/CloseableIterator.java
@@ -1,0 +1,12 @@
+package rvn;
+
+import java.io.Closeable;
+import java.util.Iterator;
+
+/**
+ *
+ * @author wozza
+ */
+public abstract class CloseableIterator<T> implements Iterator<T>, Closeable {
+    
+}
